agrego modelos y modifico restricciones

+++ src/database/migrations/20231212175632-create-cart.js
@@ -28,7 +28,7 @@ module.exports = {
         },
       },
       quantity: {
-        type: Sequelize.BIGINT,
+        type: Sequelize.INTEGER,
         allowNull: false
       },
       createdAt: {
